Support async for listener
